Tutorial 04 - nRF52 firmware version 04
This is the example for the final demo.
Added reading real sensor data, temperature, pressure, and humidity. Compatibles with myWSAN-debug4-RealSensorData.apk.
Still couldn't fix bug: 2 cluster heads may have a conflict on connecting to a same Thingy. This bug is from the orignial code of provided by Nordic, multilink.